+++ database/migrations/2020_02_26_184807_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name')->unique();
-            $table->double('price');
+            $table->float('price');
             $table->integer('creator_id');
             $table->integer('editor_id')->nullable();
             $table->nullableTimestamps();
